Changes for supporting Bridge.NET

+++ Phantasma.Cryptography/Hashing/SHA256.cs
@@ -178,7 +178,7 @@ namespace Phantasma.Cryptography.Hashing
                     amount_to_copy = 64 - pending_block_off;
                 }
 
-                Array.Copy(data, offset, pending_block, pending_block_off, amount_to_copy);
+                Array.Copy(data, (int)offset, pending_block, (int)pending_block_off, (int)amount_to_copy);
                 len -= amount_to_copy;
                 offset += amount_to_copy;
                 pending_block_off += amount_to_copy;
